# backend_monbrobroli-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "api/editCategorie/{id}": {
            "put": {
                "tags": [
                    "Abonnement."
                ],
                "summary": "Modifier la catégorie.",
                "operationId": "Modifier la catégorie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de l'emploi du temps.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Modifier la catégorie",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "example": "Etudiant"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Mauvaise méthode"
                    }
                }
            }
        },
...
        "/api/showAffiche/'mobile'ou'pc'": {
            "get": {
                "tags": [
                    "Affiches"
                ],
                "summary": "Récupère la liste des affiches filtrées par appareil",
                "description": "Retourne les affiches disponibles pour le type d'appareil envoyé (pc ou mobile)",
                "operationId": "getAffichesByAppareil",
                "parameters": [
                    {
                        "name": "appareil",
                        "in": "query",
                        "description": "Type d'appareil ('pc' ou 'mobile') pour filtrer les affiches",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "pc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des affiches récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "titre": {
                                                        "type": "string",
                                                        "example": "Affiche de bienvenue"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Ceci est une affiche test"
                                                    },
                                                    "appareil": {
                                                        "type": "string",
                                                        "example": "pc"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-10-03T12:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-10-03T12:30:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/list_offres": {
            "get": {
                "tags": [
                    "Offres."
                ],
                "summary": "Voir les offres",
                "description": "Retourne la liste des offres.",
                "operationId": "Voir les offres.",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/search_offres": {
            "post": {
                "tags": [
                    "Offres."
                ],
                "summary": "Recherche sur les offres",
                "description": "Recherche avancée sur les offres par nom, lieu et catégories",
                "operationId": "1417e7120cb04e39239de6fd0640d60e",
                "requestBody": {
                    "description": "Paramètres de recherche",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom_offre",
                                    "lieu",
                                    "categorie"
                                ],
                                "properties": {
                                    "nom_offre": {
                                        "type": "string",
                                        "example": "Serveur en restaurant"
                                    },
                                    "lieu": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "categorie": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "restauration"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}